requirements.txt
streamlit
llama-cpp-python
elevenlabs
python-dotenv
pandas